chore: move projectDiagnostics.ts
